// circabc-frontend/src/app/core/generated/circabc/model/dynamicPropertyDefinition.ts
/**
 * CIRCABC REST API
 * This is the first version of the CIRCABC REST API used by the new User Interface 
 *
 * The version of the OpenAPI document: 4.2.0
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DynamicPropertyDefinitionUpdatedValues } from './dynamicPropertyDefinitionUpdatedValues';


export interface DynamicPropertyDefinition { 
    readonly id?: string;
    readonly index?: number;
    readonly name?: string;
    /**
     * The object that is used to compile all the translations of a node into a JSON object. It is basically composed of a map with a key languaguage code and its value 
     */
    title: { [key: string]: string; };
    propertyType: DynamicPropertyDefinition.PropertyTypeEnum;
    /**
     * non mandatory field, used only when having select type 
     */
    possibleValues?: Array<string>;
    /**
     * non mandatory field, used only when having select type 
     */
    updatedValues?: Array<DynamicPropertyDefinitionUpdatedValues>;
}
export namespace DynamicPropertyDefinition {
    export type PropertyTypeEnum = 'DATE_FIELD' | 'TEXT_FIELD' | 'TEXT_AREA' | 'SELECTION' | 'MULTI_SELECTION';
    export const PropertyTypeEnum = {
        DATEFIELD: 'DATE_FIELD' as PropertyTypeEnum,
        TEXTFIELD: 'TEXT_FIELD' as PropertyTypeEnum,
        TEXTAREA: 'TEXT_AREA' as PropertyTypeEnum,
        SELECTION: 'SELECTION' as PropertyTypeEnum,
        MULTISELECTION: 'MULTI_SELECTION' as PropertyTypeEnum
    };
}